+++ classes/multilingual-plugins/easy-language/class-parser-base.php
@@ -124,10 +124,7 @@ class Parser_Base {
 	 * @return bool
 	 */
 	public function is_object_using_pagebuilder( Post_Object $post_object ): bool {
-		if ( $post_object->is_simplifiable() || $post_object->is_simplified() ) {
-			return true;
-		}
-		return false;
+		return $post_object->is_simplifiable() || $post_object->is_simplified();
 	}
 
 	/**