+++ classes/multilingual-plugins/polylang/class-init.php
@@ -105,15 +105,6 @@ class Init extends Base implements Multilingual_Plugins_Base {
 	}
 
 	/**
-	 * Return languages this plugin will support.
-	 *
-	 * @return array
-	 */
-	public function get_supported_languages(): array {
-		return array();
-	}
-
-	/**
 	 * Run on uninstallation.
 	 *
 	 * @return void