+++ classes/multilingual-plugins/translatepress/class-translatepress-summ-ai-machine-translator.php
@@ -138,7 +138,7 @@ class Translatepress_Summ_Ai_Machine_Translator extends TRP_Machine_Translator {
 			}
 		}
 
-		// will have the same indexes as $new_strings or it will be an empty array if something went wrong.
+		// will have the same indexes as $new_strings, or it will be an empty array if something went wrong.
 		return $translated_strings;
 	}