+++ uninstall.php
@@ -17,6 +17,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
+// do nothing if PHP-version is not 8.0 or newer.
+if ( PHP_VERSION_ID < 80000 ) { // @phpstan-ignore smaller.alwaysFalse
+	return;
+}
+
 // save plugin-path.
 const EASY_LANGUAGE = __FILE__;
 
@@ -24,14 +29,24 @@ const EASY_LANGUAGE = __FILE__;
 require_once __DIR__ . '/inc/autoload.php';
 require_once __DIR__ . '/inc/constants.php';
 
-// include all API-files.
-foreach ( glob( plugin_dir_path( EASY_LANGUAGE ) . 'inc/apis/*.php' ) as $filename ) {
-	require_once $filename;
+// get API directory files.
+$api_directory_files = glob( plugin_dir_path( EASY_LANGUAGE ) . 'inc/apis/*.php' );
+
+// include all API-files, if they are given.
+if( is_array( $api_directory_files ) ) {
+	foreach ( $api_directory_files as $filename ) {
+		require_once $filename;
+	}
 }
 
-// include all settings-files.
-foreach ( glob( plugin_dir_path( EASY_LANGUAGE ) . 'inc/multilingual-plugins/*.php' ) as $filename ) {
-	require_once $filename;
+// get plugin directory files.
+$plugin_directory_files = glob( plugin_dir_path( EASY_LANGUAGE ) . 'inc/multilingual-plugins/*.php' );
+
+// include all settings-files, if they are given.
+if( is_array( $plugin_directory_files ) ) {
+	foreach ( $plugin_directory_files as $filename ) {
+		require_once $filename;
+	}
 }
 
 /**